g\gygy

+++ routes/web.php
@@ -4,13 +4,20 @@ use App\Http\Controllers\CategoryController;
 use App\Http\Controllers\CityController;
 use App\Http\Controllers\CountryController;
 use App\Http\Controllers\DataImportController;
+use App\Http\Controllers\InfoShopController;
+use App\Http\Controllers\Login_con;
 use App\Http\Controllers\ProductController;
+use App\Http\Controllers\ProductShopController;
 use App\Http\Controllers\ReviewController;
 use App\Http\Controllers\RoleController;
+use App\Http\Controllers\ShopCartController;
+use App\Http\Controllers\ShopController;
+use App\Http\Controllers\ShowAdminShopController;
 use App\Http\Controllers\StateController;
 use App\Http\Controllers\TagController;
 use App\Http\Controllers\TicketController;
 use App\Http\Controllers\UnitController;
+use App\Http\Controllers\UserAuthController;
 use App\Models\City;
 use App\Models\Country;
 use App\Models\Image;
@@ -86,8 +93,64 @@ Route::get('test_tag',function (){
     $tag = User::find(502);
     return $tag->roles;
 });
+
 */
+//Route::post('login', [UserAuthController::class, 'loginUser'])->name('login')->middleware(['auth','user_is_admin']);
+//Route::view("loginUser",'loginUser');
+
+//Route::get('shop', [ProductShopController::class, 'index'])->name('shop');
+Route::get('ShopAdmin', [ShowAdminShopController::class, 'index'])->name('ShopAdmin');
+Route::get('info/{id}', [ShowAdminShopController::class, 'ShowUpdateAdmin'])->name('info');
+//shop_cart
+Route::get('Shop_Cart', [ShopCartController::class, 'index'])->name('Shop_Cart');
+Route::post('AddCartShop', [ShopCartController::class, 'store']);
+Route::get('edit-Cart/{id}', [ShopCartController::class, 'ShowUpdate'])->name('edit-Cart');
+Route::post('update', [ShopCartController::class, 'update'])->name('update');
+Route::delete('Shop_Cart', [ShopCartController::class, 'delete'])->name('delete-Cart');
+
+
+//shop_info
+Route::get('Shop', [InfoShopController::class, 'index'])->name('Shop');
+Route::get('addShop', [InfoShopController::class, 'addShopShow'])->name('addShop');
+Route::post('add_shop', [InfoShopController::class, 'store']);
+Route::get('updateShop/{id}', [InfoShopController::class, 'ShowUpdate'])->name('updateShop');
+Route::post('DoUpdateShop', [InfoShopController::class, 'DoShowUpdateShop'])->name('DoUpdateShop');
+/*Route::post('new-productShop', [ProductShopController::class, 'store']);
+Route::get('new-productShop',[ProductShopController::class ,'newProduct'])->name('new-productShop');
+Route::put('update-productShop', [ProductShopController::class, 'update'])->name('update-productShop');
+
+Route::get('update-productShop/{id?}',[ProductShopController::class ,'newProduct'])->name('update-productShop-form');
+Route::delete('new-productShop/{id}', [ProductShopController::class, 'delete']);
+Route::delete('new-productShop/{id}', [ProductShopController::class, 'deleteOption'])->name('dddd');*/
+
+
+
+
+
+
+
+//ProductShop
+
+    Route::get('productsShop', [ProductShopController::class, 'index'])->name('productsShop');
+    Route::post('new-productShop', [ProductShopController::class, 'store']);
+    Route::get('new-productShop',[ProductShopController::class ,'newProduct'])->name('new-productShop');
+    Route::put('update-productShop', [ProductShopController::class, 'update'])->name('update-productShop');
+
+    Route::get('update-productShop/{id?}',[ProductShopController::class ,'newProduct'])->name('update-productShop-form');
+    Route::delete('new-productShop/{id}', [ProductShopController::class, 'delete']);
+    Route::delete('new-productShop/{id}', [ProductShopController::class, 'deleteOption'])->name('dddd');
+
+
+
+
+
+
+
+
+
+
 Route::group(['auth'=>'user_is_admin'], function(){
+
 //Unit
     Route::get('units', [UnitController::class, 'index'])->name('units')->middleware(['auth','user_is_admin']);
     Route::post('units', [UnitController::class, 'store']);
@@ -111,7 +174,7 @@ Route::group(['auth'=>'user_is_admin'], function(){
     Route::put('categories', [CategoryController::class, 'update']);
     Route::delete('categories', [CategoryController::class, 'delete']);
     Route::get('edit-image/{id}', [CategoryController::class, 'editImage'])->name('edit-image');;
-    Route::post('updateImage', [CategoryController::class, 'updateImage'])->name('image.update');
+    Route::put('updateImage/{id}', [CategoryController::class, 'updateImage'])->name('image.update');
 
 
 
@@ -127,13 +190,12 @@ Route::group(['auth'=>'user_is_admin'], function(){
 //products
     Route::get('products', [ProductController::class, 'index'])->name('products')->middleware(['auth','user_is_admin']);
     Route::post('new-product', [ProductController::class, 'store']);
-    Route::get('new-product',[ProductController::class ,'newProduct'])->name('new-product');
-    Route::put('update-product', [ProductController::class, 'update'])->name('update-product');
+    Route::get('new-product',[ProductController::class ,'newProduct'])->name('new-product')->middleware(['auth','user_is_admin']);
+    Route::put('update-product', [ProductController::class, 'update'])->name('update-product')->middleware(['auth','user_is_admin']);
 
-    Route::get('update-product/{id?}',[ProductController::class ,'newProduct'])->name('update-product-form');
+    Route::get('update-product/{id?}',[ProductController::class ,'newProduct'])->name('update-product-form')->middleware(['auth','user_is_admin']);
     Route::delete('new-product/{id}', [ProductController::class, 'delete']);
-    Route::delete('new-product/{id}', [ProductController::class, 'deleteOption'])->name('dddd');
-
+    Route::delete('new-product/{id}', [ProductController::class, 'deleteOption'])->name('dddd')->middleware(['auth','user_is_admin']);
